backend/ingest-api/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pymupdf
pytesseract
python-docx
pypandoc
openpyxl
python-pptx
Pillow
pandas
beautifulsoup4
markdownify
watchdog
chromadb==0.5.11
pyyaml
prometheus-client
python-jose[cryptography]
httpx
nltk
